api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.6.3
python-dotenv==1.0.1
razorpay==1.4.1
python-jose[cryptography]==3.3.0